+++ tests/catalog/test_manage.py
@@ -1,7 +1,10 @@
+import json
 import sys
 
 import dirty_equals
 import pytest
+
+from openeogeotrellis.catalog.enrich import LinksList
 
 if sys.version_info < (3, 10):
     # TODO #1060 clean up once python 3.8/3.9 support can be dropped
@@ -16,7 +19,31 @@ from openeogeotrellis.catalog.manage import (
     ENRICHMENT_MODE,
     extract_band_metadata_list,
     apply_raster_scale_and_offset_to_band_metadata,
+    LayerCatalog,
+    sort_dict_like_other,
 )
+
+
+class TestLayerCatalog:
+    def test_basic(self, tmp_path):
+        layer_catalog = LayerCatalog()
+        layer_catalog.set_collection_metadata(
+            {
+                "id": "FOO",
+                "title": "The FOO dataset",
+            }
+        )
+        path = tmp_path / "layercatalog.json"
+        layer_catalog.write_json_file(path)
+
+        assert json.loads(path.read_text(encoding="utf-8")) == [{"id": "FOO", "title": "The FOO dataset"}]
+
+    def test_compact_settings(self, tmp_path):
+        layer_catalog = LayerCatalog()
+        layer_catalog.set_collection_metadata({"id": "FOO", "title": "The FOO dataset"})
+        path = tmp_path / "layercatalog.json"
+        layer_catalog.write_json_file(path, indent=None, separators=(",", ":"))
+        assert path.read_text(encoding="utf-8") == '[{"id":"FOO","title":"The FOO dataset"}]\n'
 
 
 class TestBandMetadata:
@@ -98,7 +125,18 @@ class TestBuildMetadata:
                     {
                         "rel": "root",
                         "href": "https://stac.test/",
-                    }
+                        "type": "application/json",
+                    },
+                    {
+                        "rel": "self",
+                        "href": "https://stac.test/c/foobar1",
+                        "type": "application/json",
+                    },
+                    {
+                        "rel": "about",
+                        "href": "https://stac.test/c/foobar1.html",
+                        "type": "text/html",
+                    },
                 ],
             ),
         )
@@ -181,7 +219,11 @@ class TestBuildMetadata:
                 "eo:bands": [{"name": "blue", "description": "Not red"}],
                 "raster:bands": [{"name": "blue"}],
             },
-            "links": [{"href": "https://stac.test/", "rel": "root"}],
+            "links": [
+                {"rel": "root", "href": "https://stac.test/", "type": "application/json"},
+                {"rel": "self", "href": "https://stac.test/c/foobar1", "type": "application/json"},
+                {"rel": "about", "href": "https://stac.test/c/foobar1.html", "type": "text/html"},
+            ],
         }
 
     def test_bands_metadata(self):
@@ -249,6 +291,26 @@ class TestBuildMetadata:
                         }
                     ],
                 },
+            }
+        )
+
+    def test_enrich_at_build_time_with_upstream_links_filter(self):
+        def upstream_links_filter(links: LinksList) -> LinksList:
+            return [link for link in links if link.get("rel") in {"license", "about"}]
+
+        metadata = build_stac_collection_metadata(
+            id="FOOBAR",
+            stac_url="https://stac.test/c/foobar1",
+            bands=[BandMetadata(name="blue", description="Not red")],
+            enrichment_mode=ENRICHMENT_MODE.LEGACY_AT_BUILD_TIME,
+            upstream_links_filter=upstream_links_filter,
+        )
+        assert metadata == dirty_equals.IsPartialDict(
+            {
+                "id": "FOOBAR",
+                "links": [
+                    {"rel": "about", "href": "https://stac.test/c/foobar1.html", "type": "text/html"},
+                ],
             }
         )
 
@@ -543,8 +605,8 @@ class TestBandMetadataList:
             BandMetadata(name="bar", data_type="uint16", raster_scale=0.001),
         ]
         assert apply_raster_scale_and_offset_to_band_metadata(bands) == [
-            BandMetadata(name="foo", data_type="float64"),
-            BandMetadata(name="bar", data_type="float64"),
+            BandMetadata(name="foo", data_type="float32"),
+            BandMetadata(name="bar", data_type="float32"),
         ]
 
     def test_apply_raster_scale_and_offset_with_offset(self):
@@ -553,8 +615,8 @@ class TestBandMetadataList:
             BandMetadata(name="bar", data_type="uint16", raster_offset=-1.5),
         ]
         assert apply_raster_scale_and_offset_to_band_metadata(bands) == [
-            BandMetadata(name="foo", data_type="float64"),
-            BandMetadata(name="bar", data_type="float64"),
+            BandMetadata(name="foo", data_type="float32"),
+            BandMetadata(name="bar", data_type="float32"),
         ]
 
     def test_apply_raster_scale_and_offset_with_classification_classes(self):
@@ -570,5 +632,36 @@ class TestBandMetadataList:
             ),
         ]
         assert apply_raster_scale_and_offset_to_band_metadata(bands) == [
-            BandMetadata(name="foo", data_type="float64"),
+            BandMetadata(name="foo", data_type="float32"),
         ]
+
+
+@pytest.mark.parametrize(
+    ["other", "expected"],
+    [
+        (
+            {},
+            '{"id": "foo", "description": "The foo", "bands": [1, 2, 3], "link": true}',
+        ),
+        (
+            {"id": "bar", "bands": 3},
+            '{"id": "foo", "bands": [1, 2, 3], "description": "The foo", "link": true}',
+        ),
+        (
+            ["id", "color", "link", "flavor"],
+            '{"id": "foo", "link": true, "description": "The foo", "bands": [1, 2, 3]}',
+        ),
+    ],
+)
+def test_sort_dict_like_other(other, expected):
+    orig = {
+        "id": "foo",
+        "description": "The foo",
+        "bands": [1, 2, 3],
+        "link": True,
+    }
+    assert json.dumps(orig) == '{"id": "foo", "description": "The foo", "bands": [1, 2, 3], "link": true}'
+
+    reordered = sort_dict_like_other(orig, other)
+    assert reordered == orig
+    assert json.dumps(reordered) == expected