+++ tests/test_layercatalog.py
@@ -1,5 +1,8 @@
+import gzip
+import json
 import logging
 import unittest.mock as mock
+import zipfile
 from pathlib import Path
 from typing import List, Tuple, Union
 
@@ -23,6 +26,7 @@ from openeogeotrellis.layercatalog import (
     _get_sar_backscatter_arguments,
     _merge_layers_with_common_name,
     get_layer_catalog,
+    _read_catalog_file,
 )
 from openeogeotrellis.testing import gps_config_overrides
 from openeogeotrellis.vault import Vault
@@ -269,6 +273,70 @@ def test_layer_catalog_step_resolution(vault):
                 warnings += warn_str + "\n"
                 break
     assert warnings == ""
+
+
+@pytest.fixture
+def layercatalog_json_gz(tmp_path) -> Path:
+    """Fixture for building a layercatalog.json.gz file"""
+    layer_catalog_data = [
+        {"id": "ZOO", "description": "The ZOO layer"},
+        {"id": "ZAZ", "description": "The ZAZ layer"},
+    ]
+    path = tmp_path / "layercatalog01.json.gz"
+    assert not path.exists()
+
+    with gzip.open(path, mode="wt") as f:
+        json.dump(obj=layer_catalog_data, fp=f, indent=2)
+    # Check GZIP signature
+    assert path.read_bytes()[:2] == b"\x1f\x8b"
+
+    return path
+
+
+def test_get_layer_catalog_from_gzip(layercatalog_json_gz):
+    with gps_config_overrides(layer_catalog_files=[str(layercatalog_json_gz)]):
+        catalog = get_layer_catalog()
+
+    assert sorted(c["id"] for c in catalog.get_all_metadata()) == ["ZAZ", "ZOO"]
+
+
+def test_read_catalog_file_json():
+    path = get_test_data_file("layercatalog01.json")
+    assert _read_catalog_file(path) == {
+        "BAR": {"id": "BAR", "description": "bar", "links": ["example.com/bar"]},
+        "BZZ": {"id": "BZZ"},
+        "FOO": {"id": "FOO", "license": "mit"},
+    }
+
+
+def test_read_catalog_file_json_gz(layercatalog_json_gz):
+    assert _read_catalog_file(layercatalog_json_gz) == {
+        "ZAZ": {"id": "ZAZ", "description": "The ZAZ layer"},
+        "ZOO": {"id": "ZOO", "description": "The ZOO layer"},
+    }
+
+
+def test_read_catalog_file_zip(tmp_path):
+    layercatalog_zip = tmp_path / "layercatalog.zip"
+    with zipfile.ZipFile(layercatalog_zip, mode="w") as zf:
+        # A file with single, top-level collection (no list)
+        zf.writestr(
+            "one-collection.json",
+            json.dumps({"id": "THE_ONE"}),
+        )
+        # A file with list of collections
+        zf.writestr(
+            "multiple.json",
+            json.dumps([{"id": "TWO"}, {"id": "THREE"}]),
+        )
+    # Check ZIP signature
+    assert layercatalog_zip.read_bytes()[:2] == b"\x50\x4b"
+
+    assert _read_catalog_file(layercatalog_zip) == {
+        "THE_ONE": {"id": "THE_ONE"},
+        "TWO": {"id": "TWO"},
+        "THREE": {"id": "THREE"},
+    }
 
 
 def test_merge_layers_with_common_name_nothing():