+++ tests/test_load_stac.py
@@ -4163,15 +4163,31 @@ class TestPrepareContext:
         ]
 
     @pytest.mark.parametrize(
-        ["aggregate_spatial_geometries", "expected_items"],
+        ["aggregate_spatial_geometries", "feature_flags", "env_dict", "expected_items"],
         [
             (
                 shapely.geometry.box(2.5, 49.5, 3.5, 51.5),
+                {},
+                {},
                 ["item-1", "item-2"],
             ),
             (
                 shapely.geometry.Polygon([(6.2, 49.8), (6.5, 51.5), (7, 49), (2.5, 49.5), (6.2, 49.8)]),
+                {},
+                {},
                 ["item-1", "item-3"],
+            ),
+            (
+                shapely.geometry.Polygon([(6.2, 49.8), (6.5, 51.5), (7, 49), (2.5, 49.5), (6.2, 49.8)]),
+                {"stac_api_filter_by_geometry": False},
+                {},
+                ["item-1", "item-2", "item-3"],
+            ),
+            (
+                shapely.geometry.Polygon([(6.2, 49.8), (6.5, 51.5), (7, 49), (2.5, 49.5), (6.2, 49.8)]),
+                {},
+                {"stac_api_filter_by_geometry": False},
+                ["item-1", "item-2", "item-3"],
             ),
             (
                 DriverVectorCube(
@@ -4182,12 +4198,14 @@ class TestPrepareContext:
                         ]
                     )
                 ),
+                {},
+                {},
                 ["item-1", "item-3"],
             ),
         ],
     )
     def test_prepare_context_spatial_filtering_geometries(
-        self, dummy_stac_api_server, aggregate_spatial_geometries, expected_items
+        self, dummy_stac_api_server, aggregate_spatial_geometries, expected_items, feature_flags, env_dict
     ):
         with dummy_stac_api_server.serve() as dummy_stac_api:
             context = _prepare_context(
@@ -4195,10 +4213,8 @@ class TestPrepareContext:
                 load_params=LoadParameters(
                     aggregate_spatial_geometries=aggregate_spatial_geometries,
                 ),
-                feature_flags={
-                    "stac_api_filter_by_geometry": True,
-                },
-                env=EvalEnv(),
+                feature_flags=feature_flags,
+                env=EvalEnv(env_dict),
             )
 
         dumper = OpenSearchClientDumper()